shaping: move fuzz test to guarded file
This commit moves the fuzz test to a file that only
builds on go 1.18+, allowing CI to pass on 1.17.
